+ Refactor source code
