+++ src/components/GDatePicker/Header/GDatePickerHeader.scss
@@ -1,4 +1,4 @@
-@import "var";
+@import "../var";
 
 .g-date-picker-header {
   padding: 4px 8px;
