+++ src/components/GDatePicker/Table/GDatePickerTable.scss
@@ -1,4 +1,4 @@
-@import 'var';
+@import '../var';
 
 $highlightedItem: rgb(231, 215, 253);
