+++ src/components/GDatePicker/Table/date-picker-table.js
