+++ src/components/GDatePicker/Title/GDatePickerTitle.vue
@@ -13,7 +13,8 @@
 
 <script>
   import { computed } from '@vue/composition-api'
-  import { TRANSITION_NAMES } from './utils';
+  import { TRANSITION_NAMES } from '../utils';
+  import { getYearButtonClickHandler } from './GDatePickerTitleUtil'
 
   // class
   const PICKER_TITLE = 'g-date-picker-title'
@@ -24,29 +25,28 @@
   const PICKER_TITLE_YEAR = 'g-date-picker-title__year'
   const PICKER_TITLE_DATE = 'g-date-picker-title__date'
 
-
-  // exposed events
-  export const EVENT_NAMES = {
-    UPDATE_SELECTING_YEAR: 'update:selecting-year'
-  }
-
   export default {
     name: 'GDatePickerTitle',
     props: {
+      // 1st line -> year
       year: {
         type: [Number, String],
         default: ''
-      }, // 1st line -> year
+      },
+      // 2nd line, maybe contain html code -> render as raw html
       date: {
         type: String,
         default: ''
-      },// 2nd line, maybe contain html code -> render as raw html
-      disabled: Boolean,      // Boolean value indicate that this component is disabled or not
-      readonly: Boolean,      // Boolean value indicate that this component is readonly or not
-      selectingYear: Boolean, // Boolean value indicate that whether Years is showing or not
+      },
+      // Boolean value indicate that this component is disabled or not
+      disabled: Boolean,
+      // Boolean value indicate that this component is readonly or not
+      readonly: Boolean,
+      // Boolean value indicate that whether Years is showing or not
+      selectingYear: Boolean,
     },
     setup(props, context) {
-      let titleClasses = computed((() => {
+      const titleClasses = computed((() => {
         return {
           [PICKER_TITLE]: true,
           [PICKER_TITLE_DISABLED]: props.disabled
@@ -54,7 +54,7 @@
       }))
 
       // Year info
-      let titleYearClasses = computed((() => {
+      const titleYearClasses = computed((() => {
         return {
           [PICKER_TITLE_BTN]: true,
           [PICKER_TITLE_BTN_READONLY]: props.readonly,
@@ -62,11 +62,8 @@
           [PICKER_TITLE_YEAR]: true
         }
       }))
-      function onYearButtonClicked() {
-        if (props.selectingYear || props.readonly)
-          return
-        context.emit(EVENT_NAMES.UPDATE_SELECTING_YEAR, true)
-      }
+
+      const onYearButtonClicked = getYearButtonClickHandler(props, context)
 
       // Date/selected items state
       let titleDateClasses = computed(() => {
