+++ src/components/GDatePicker/dateFilter.js
@@ -1,24 +1,25 @@
 /**
  * Check if provided value is in range
+ * @private
+ * @publicForTestOnly
  * @param value
  * @param min
  * @param max
- * @param isValueInRangeCustomFn
  * @returns {boolean|*}
  */
-export const isValueInRange = (value, min, max) => {
+export const _isValueInRange = (value, min, max) => {
   return (!min || value >= min) && (!max || value <= max)
 }
 
-export function isDateAllowed(value, filterFn) {
+/**
+ * @private
+ * @publicForTestOnly
+ */
+export function _isDateAllowed(value, filterFn) {
   if (typeof filterFn !== 'function')
     return true
 
   return filterFn(value)
 }
 
-export function dateFilter(props) {
-  return function isDateAvailable(value) {
-    return isDateAllowed(value, props.allowedDates) && isValueInRange(value, props.min, props.max)
-  }
-}
+export default (props) => (value) => _isDateAllowed(value, props.allowedDates) && _isValueInRange(value, props.min, props.max)
